fixed possible crashes on exception at the end of segment user command's help opens DC++ webhelp updated zlib to 1.2.4 migrated to MSVS2010 added option to disable Explorer theming in listviews and treeviews added temporary workaround for possible DHT search spam first try to fix throttling with very large speeds updated boost to 1.43

+++ trunk/client/ThrottleManager.h
@@ -51,24 +51,24 @@ namespace dcpp
 		/*
 		 * Returns current download limit.
 		 */
-		int64_t getDownloadLimit() const;
+		size_t getDownloadLimit() const;
 
 		/*
 		 * Returns current download limit.
 		 */
-		int64_t getUploadLimit() const;
+		size_t getUploadLimit() const;
 		
 	private:
 		
 		// download limiter
-		int							downLimit;
-		int64_t						downTokens;
+		size_t						downLimit;
+		size_t						downTokens;
 		boost::condition_variable	downCond;
 		boost::mutex				downMutex;
 		
 		// upload limiter
-		int							upLimit;
-		int64_t						upTokens;
+		size_t						upLimit;
+		size_t						upTokens;
 		boost::condition_variable	upCond;
 		boost::mutex				upMutex;
